upti-k

+++ app/Views/data_akademik/jurusan.php
@@ -29,7 +29,7 @@
             // Hitung nilai $i
             $i = 1 + (10 * ($page - 1));
             ?>
-            <?php foreach ($jurusan as $ta): ?>
+            <?php foreach ($jurusan as $ta) : ?>
                 <tr class="text-center">
                     <td scope="row-1">
                         <?= $i++; ?>
@@ -38,9 +38,8 @@
                         <?= $ta['nama_jurusan']; ?>
                     </td>
                     <td>
+                        <a href="/admin/hapus_data_jurusan/<?= $ta['id_jurusan']; ?>" onclick="return confirm('Apakah anda yakin ingin menghapus data ini?');" class="btn btn-danger">Delete</a>
                         <a href="/admin/edit_data_jurusan/<?= $ta['id_jurusan']; ?>" class="btn btn-success">Edit</a>
-                        <a href="/admin/hapus_data_jurusan/<?= $ta['id_jurusan']; ?>"
-                            onclick="return confirm('apakah anda yakin');" class="btn btn-danger">Delete</a>
                     </td>
 
 
@@ -67,7 +66,6 @@
 
 <?= $this->section('scripts'); ?>
 <script>
-
     const swalElement = document.querySelector('.swal'); // Mengambil elemen dengan kelas '.swal'
     const swalData = swalElement.dataset.swal; // Mengambil data dari atribut data HTML 'data-swal'
