+++ app/Views/data_akademik/prodi.php
@@ -14,8 +14,8 @@
                 <th scope="col">No</th>
                 <th scope="col">Kode Prodi</th>
                 <th scope="col">Jurusan</th>
-                <th scope="col">Nama Prodi</th>
                 <th scope="col">Program</th>
+                <th scope="col">Nama Program Studi</th>
                 <th scope="col">Aksi</th>
 
 
@@ -32,7 +32,7 @@
             // Hitung nilai $i
             $i = 1 + (10 * ($page - 1));
             ?>
-            <?php foreach ($prodi as $ta): ?>
+            <?php foreach ($prodi as $ta) : ?>
                 <tr class="text-center">
                     <td scope="row-1">
                         <?= $i++; ?>
@@ -50,9 +50,8 @@
                         <?= $ta['nama_prodi']; ?>
                     </td>
                     <td>
+                        <a href="/admin/hapus_data_prodi/<?= $ta['id_prodi']; ?>" onclick="return confirm('Apakah anda yakin ingin menghapus data ini?');" class="btn btn-danger">Delete</a>
                         <a href="/admin/edit_data_prodi/<?= $ta['id_prodi']; ?>" class="btn btn-success">Edit</a>
-                        <a href="/admin/hapus_data_prodi/<?= $ta['id_prodi']; ?>"
-                            onclick="return confirm('apakah anda yakin');" class="btn btn-danger">Delete</a>
                     </td>
 
 
@@ -63,7 +62,7 @@
 
     <div class="row my-3">
         <div class="col">
-            <a href="<?= route_to('admin.add.prodi') ?>" class="btn btn-primary">Tambah Data prodi</a>
+            <a href="<?= route_to('admin.add.prodi') ?>" class="btn btn-primary">Tambah Data Prodi</a>
             <a href="<?= route_to('prodi.export.pdf') ?>" class="btn btn-warning">Export PDF</a>
         </div>
     </div>
@@ -79,7 +78,6 @@
 
 <?= $this->section('scripts'); ?>
 <script>
-
     const swalElement = document.querySelector('.swal'); // Mengambil elemen dengan kelas '.swal'
     const swalData = swalElement.dataset.swal; // Mengambil data dari atribut data HTML 'data-swal'
